3D viewing: cross-check fixes after self-review
Fixes 7 real issues identified by reading the prior commit's code:

1. Routing collision — moved Models list into `app/(tabs)/models.tsx` and
   removed `app/models/index.tsx`. Both files mapped to `/models`; the
   detail route at `app/models/[id].tsx` still resolves to `/models/<id>`.

2. ApsModelDerivativeConverter dropped `using` on the IHttpClientFactory-
   issued HttpClient. Disposing factory clients closes the pooled handler
   ahead of schedule, defeating connection pooling.

3. RevitGltfExporter swapped deprecated `ElementId.IntegerValue` for
   `ElementId.Value` (Int64) per the Revit 2024+ migration noted in
   CLAUDE.md.

4. RevitGltfExporter dropped the broken mid-polymesh split (triangles
   straddled the old/new node and produced invalid indices). Now uses
   32-bit indices (componentType 5125), so the 65k-vertex limit is gone
   and the split logic is unnecessary.

5. viewer-extras auto-LOD: rebuilt the sampler around requestAnimationFrame
   instead of setInterval. The old version measured the timer's own
   cadence (always ~20fps from a 50ms interval) and reliably tripped the
   downgrade after 3 seconds regardless of real render cost.

6. Dockerfile IfcConvert install is now opt-in via build-arg
   `IFCCONVERT_URL=<full-url>` (default "skip"). The previous URL pattern
   `IfcConvert-v0.7.0-linux64.zip` doesn't exist on the IfcOpenShell
   builds bucket — real filenames carry a build-hash. Robust unzip via
   find+mv handles arbitrary subfolder layouts; install verifies with
   `IfcConvert --version` so a wrong URL fails the build cleanly.

7. viewer.html handleCommand wrapper hardened against missing payloads
   (`const p = (cmd && cmd.payload) || {}`) so a stray RN message during
   boot can't TypeError out of the message loop.

Caveats still open (need Revit / device verification):
- OnFaceBegin returning Proceed (vs Skip) — uncertain CustomExporter
  convention; needs Revit testing.
- Walkthrough Y-up assumption may be wrong for Z-up Revit-exported GLBs.
- Thumbnail bounds union over all VEC3 accessors includes NORMAL (range
  -1..1) — dwarfed by POSITION in practice but not strictly correct.
