+++ Planscape.Server/docker/Dockerfile
@@ -22,17 +22,25 @@ WORKDIR /app
 
 RUN apt-get update && apt-get install -y curl && rm -rf /var/lib/apt/lists/*
 
-# P7 — IfcConvert (IfcOpenShell) for IFC→glTF derivatives. Opt-in: only used
-# when ModelConverter:Provider=ifcconvert. Pinned to a known-good release.
-ARG IFCCONVERT_VERSION=0.7.0
-RUN if [ "$IFCCONVERT_VERSION" != "skip" ]; then \
+# P7 — IfcConvert (IfcOpenShell) for IFC→glTF derivatives. OPT-IN: pass
+# `--build-arg IFCCONVERT_URL=<full-url>` to install at build time. Default
+# is "skip" because the IfcOpenShell build bucket filenames include a
+# build-hash suffix that needs to be looked up per release (e.g.
+# `IfcConvert-v0.8.0-1b1fd1e-linux64.zip`). When the URL is provided the
+# binary is unzipped to a temp dir, the IfcConvert executable is moved to
+# /usr/local/bin/IfcConvert, and the temp dir cleaned up.
+ARG IFCCONVERT_URL=skip
+RUN if [ "$IFCCONVERT_URL" != "skip" ]; then \
         apt-get update && apt-get install -y --no-install-recommends \
             ca-certificates wget unzip libxi6 libxrender1 libglu1-mesa && \
-        wget -q "https://s3.amazonaws.com/ifcopenshell-builds/IfcConvert-v${IFCCONVERT_VERSION}-linux64.zip" -O /tmp/ifcconvert.zip && \
-        unzip /tmp/ifcconvert.zip -d /usr/local/bin && \
+        mkdir -p /tmp/ifcconvert && \
+        wget -q "$IFCCONVERT_URL" -O /tmp/ifcconvert.zip && \
+        unzip -o /tmp/ifcconvert.zip -d /tmp/ifcconvert && \
+        find /tmp/ifcconvert -type f -name 'IfcConvert' -exec mv {} /usr/local/bin/IfcConvert \; && \
         chmod +x /usr/local/bin/IfcConvert && \
-        rm /tmp/ifcconvert.zip && \
-        rm -rf /var/lib/apt/lists/* ; \
+        rm -rf /tmp/ifcconvert /tmp/ifcconvert.zip && \
+        rm -rf /var/lib/apt/lists/* && \
+        /usr/local/bin/IfcConvert --version || (echo "IfcConvert install verification failed" && exit 1) ; \
     fi
 
 COPY --from=build /app .
